server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx>=0.27.0
python-multipart>=0.0.6
